qimsc now converts to double as needed

+++ matlab/qimsc.m
@@ -68,7 +68,7 @@ nanc = qp_data.info(idx).lut_nan;
 qp_data.info(idx).clim = [c0 c1];
 
 [N C] = size(lut);
-data = floor(1+(N-.0001)*(data-c0)/(c1-c0)); % normalize to color range
+data = floor(1+(N-.0001)*double(data-c0)/(c1-c0)); % normalize to color range
 data(data<1)=1;
 data(data>N)=N;
